chore(components): move ProgressPie to components
